feat: added compound ring temperature sensor
